shells/sandbox-controller/src/requirements.txt
cloudshell-shell-core>=5.0.3,<6.0.0
retrying